backend_restructured/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.1
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data validation
pydantic==2.6.1
pydantic-settings==2.2.1

email-validator==2.1.0.post1

# HTTP client
httpx==0.27.0

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.1.1
isort==5.13.2
flake8==7.0.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Utilities
python-dotenv==1.0.1
slowapi 
openai 
sendgrid 
websockets 
reportlab 
